requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27.0
e2b-code-interpreter>=1.0.0
aiobotocore>=2.12.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
